requirements.txt
numpy
sounddevice
soundfile
pynput
requests
certifi
python-dotenv
